widths: extend the sweep to parameters, and fix the three it proves
With method_t keyed by imp and attributed through the checklist, the width
check now covers every parameter as well as the return. The RBMusicARView
UpdateScore: false positive from the selector-keyed prototype is gone, which
is what the keying fix was meant to achieve.

Three real defects, each confirmed against the disassembly before changing:

StorePackMusicView -setBG: took a BOOL where the encoding is i. The binary
clamps to [0, 1] with signed gt and lt tests and the lower clamp cannot fire
for a BOOL, so the parameter is a signed int. The file's own comments already
described it as clamping a range, contradicting the declared type.

RBCustomSelectView -getCollectionViewMargin and -getCollectionViewStartY:
both returned CGFloat, a double, where the encodings are f, and both bodies
use single-precision registers throughout. Both had already been read and
verified by hand earlier today: the constants were checked and correct, and
the return width was simply not what that pass was looking at.

Twelve reports remain, all signedness rather than size, listed in the audit.
Each needs its own read before being touched.

+++ Project/RBCustomSelectView.h
@@ -48,14 +48,14 @@ NS_ASSUME_NONNULL_BEGIN
  * @return The first grid's top offset within the scroll view.
  * @ghidraAddress 0x168850
  */
-- (CGFloat)getCollectionViewStartY:(RBUserSettingDataTheme)thema;
+- (float)getCollectionViewStartY:(RBUserSettingDataTheme)thema;
 
 /**
  * @brief The vertical gap left between consecutive category grids, chosen by the iPad idiom.
  * @return The inter-grid margin.
  * @ghidraAddress 0x16889c
  */
-- (CGFloat)getCollectionViewMargin;
+- (float)getCollectionViewMargin;
 
 /**
  * @brief Reload every category grid's item content.
